Move GetWorldResponse to server.
