+++ cardinal/server/handler/query.go
@@ -5,7 +5,7 @@ import (
 	"github.com/rotisserie/eris"
 
 	servertypes "pkg.world.dev/world-engine/cardinal/server/types"
-	"pkg.world.dev/world-engine/cardinal/types/engine"
+	"pkg.world.dev/world-engine/cardinal/types"
 )
 
 // PostQuery godoc
@@ -24,7 +24,7 @@ func PostQuery(world servertypes.ProviderWorld) func(*fiber.Ctx) error {
 	return func(ctx *fiber.Ctx) error {
 		ctx.Set("Content-Type", "application/json")
 		resBz, err := world.QueryHandler(ctx.Params("name"), ctx.Body())
-		if eris.Is(err, &engine.QueryNotFoundError{}) {
+		if eris.Is(err, &types.QueryNotFoundError{}) {
 			return fiber.NewError(fiber.StatusNotFound, "query not found")
 		} else if err != nil {
 			return fiber.NewError(fiber.StatusBadRequest, "encountered an error in query: "+err.Error())
