+++ cardinal/types/info.go
@@ -1,12 +1,5 @@
 package types
 
-type GetWorldResponse struct {
-	Namespace  string        `json:"namespace"`
-	Components []FieldDetail `json:"components"` // list of component names
-	Messages   []FieldDetail `json:"messages"`
-	Queries    []FieldDetail `json:"queries"`
-}
-
 // FieldDetail represents a field from a url request.
 type FieldDetail struct {
 	Name   string         `json:"name"`   // name of the message or query
