+++ cardinal/types/query.go
@@ -1,4 +1,4 @@
-package engine
+package types
 
 import "fmt"
